+++ saqc/core/modules/breaks.py
@@ -23,7 +23,7 @@ class Breaks:
         gap_window: str,
         group_window: str,
         flag: float = BAD,
-        **kwargs
+        **kwargs,
     ) -> saqc.SaQC:
         return self._defer("flagIsolated", locals())
 
@@ -34,6 +34,6 @@ class Breaks:
         window: str,
         min_periods: int = 1,
         flag: float = BAD,
-        **kwargs
+        **kwargs,
     ) -> saqc.SaQC:
         return self._defer("flagJumps", locals())