+++ saqc/core/modules/constants.py
@@ -20,7 +20,7 @@ class Constants:
         maxna: int = None,
         maxna_group: int = None,
         flag: float = BAD,
-        **kwargs
+        **kwargs,
     ) -> saqc.SaQC:
         return self._defer("flagByVariance", locals())